requirements.txt
# Backend LabelDiff - Azure App Service
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0

# Auth
msal>=1.31.0
requests>=2.32.0
PyJWT>=2.10.0

# Azure Storage
azure-storage-blob>=12.23.0

# Pydantic
pydantic>=2.0.0
